Replace rsa macro functions with real functions
Use the function signature described in the man page.

With the macro version, there where at least two issues.

1)
User code would emit warnings when compiling with `-Wcast-qual` on
GCC, since `(void*)` would cast `const` away. This should be an
implementation detail.
The cast is necessary as `EVP_PKEY_CTX_ctrl` takes a `void*` and not a
`const void*`.

2)
The unconditional cast to `void*` of the last parameter permits to pass
any parameter type to the macros without triggering any warning, making
the functions easy to misuse and possibly triggering undefined behaviour.